overflow issue in town-generator.ipynb under np.int16 default fixed

+++ sim/lib/settings/town_settings_bern.py
@@ -24,14 +24,14 @@ population_per_age_group = np.array([100744,     # 0-9
                                      124107,     # 60-69
                                      96542,      # 70-79
                                      60710       # 80+
-                                     ])
+                                     ], dtype=np.int32)
 
 region_population = population_per_age_group.sum()
 town_population = 133791
 
 # Roughly 5k tests per day in Switzerland (rough average over time frame 10.03.-27.04.2020:
 # https://www.bag.admin.ch/bag/en/home/krankheiten/ausbrueche-epidemien-pandemien/aktuelle-ausbrueche-epidemien/novel-cov/situation-schweiz-und-international.html
-daily_tests_unscaled = int(5000 * town_population / 8570000)
+daily_tests_unscaled = int(5000 * (town_population / 8570000))
 
 # Information about household structure (set to None if not available)
 # Source for Switzerland: https://www.bfs.admin.ch/bfs/de/home/statistiken/bevoelkerung/stand-entwicklung/haushalte.html
