+++ sim/lib/settings/town_settings_kaiserslautern.py
@@ -21,13 +21,13 @@ population_per_age_group = np.array([
     33065,  # 15-34
     31752,  # 35-59
     20545,  # 60-79
-    6071])  # 80+
+    6071], dtype=np.int32)  # 80+
 
 region_population = population_per_age_group.sum()
 town_population = region_population 
 
 # Roughly 100k tests per day in Germany: https://www.rki.de/DE/Content/Infekt/EpidBull/Archiv/2020/Ausgaben/15_20.pdf?__blob=publicationFile
-daily_tests_unscaled = int(100000 * town_population / 83000000)
+daily_tests_unscaled = int(100000 * (town_population / 83000000))
 
 # Information about household structure (set to None if not available)
 # Source for Germany: https://www.destatis.de/EN/Themes/Society-Environment/Population/Households-Families/Tables/lrbev05.html
